requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
Babel==2.12.1
boto==2.49.0
boto3==1.16.49
botocore==1.19.49
Django==3.1.5
django-cors-headers==3.6.0
django-money==3.0.0
django-storages==1.11.1
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
Pillow==9.4.0
promise==2.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
py-moneyed==2.0
PyJWT==2.0.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.2
